+++ src/all.ts
@@ -9,6 +9,7 @@ export { CpsButtonGroup } from './components/button-group.js';
 export { CpsCard } from './components/card.js';
 export { CpsCheckbox } from './components/checkbox.js';
 export { CpsChip } from './components/chip.js';
+export { CpsDialog } from './components/dialog.js';
 export { CpsDropdown } from './components/dropdown.js';
 export { CpsFlyout } from './components/flyout.js';
 export { CpsFormatBytes } from './components/format-bytes.js';
@@ -44,9 +45,9 @@ export { CpsTooltip } from './components/tooltip.js';
 // Utilities
 export * from './utilities/animation.js';
 export * from './utilities/base-path.js';
+export * from './utilities/form.js';
 export * from './utilities/icon-library.js';
 export * from './utilities/theme.js';
-export * from './utilities/form.js';
 
 // Events
 export * from './events/index.js';